requirements.txt
feedparser==6.0.12
httpx==0.28.1
anthropic==0.50.0
python-telegram-bot==22.7
APScheduler==3.11.2
SQLAlchemy==2.0.49
PyYAML==6.0.3
python-dotenv==1.2.2
